fix #221
